cent mult - 2

+++ MpdFlowEventPlane.cxx
@@ -552,7 +552,7 @@ float MpdFlowEventPlane::GetCentValerii(MpdAnalysisEvent &event){
    // Reject empty events (UrQMD, PHSD)
    //mMCTracks = event.fMCTrack;
    mMpdGlobalTracks = event.fMPDEvent->GetGlobalTracks();
-   float _Mult = mMpdGlobalTracks->GetEntriesFast();
+   long int _Mult = mMpdGlobalTracks->GetEntriesFast();
    /*
    int nTrMc = 0;
    for (int i = 0; i < mMCTracks->GetEntriesFast(); i++) {
@@ -598,7 +598,7 @@ float MpdFlowEventPlane::GetCentValerii(MpdAnalysisEvent &event){
    }
    */
    for(int i=0; i<14; i++){
-      if (_Mult >= (float)fMinMult_Valerii[i] && _Mult < (float)fMaxMult_Valerii[i]){
+      if ( (float)_Mult >= (float)fMinMult_Valerii[i] && (float)_Mult < (float)fMaxMult_Valerii[i]){
          return ( fMinCentPercent_Valerii[i] + fMaxCentPercent_Valerii[i] )/2.;
       }
    }
